% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{alpha.points.volcano}
\alias{alpha.points.volcano}
\title{alpha points for volcano plot}
\usage{
alpha.points.volcano(
  input.data,
  highlight.genes = goi,
  baseline.alpha = 0.1,
  highlight.alpha = 1
)
}
\arguments{
\item{input.data}{A data frame containing the expression data.}

\item{highlight.genes}{A vector containing the gene names of interest.}

\item{baseline.alpha}{The alpha value for the genes that are not of interest.}

\item{highlight.alpha}{The alpha value for the genes of interest.}
}
\value{
A vector containing the alpha value for each gene.
}
\description{
This function is used to set the alpha value for each gene in the plot.
}
\examples{

# "EnhancedVolcano" package from Github repo
# devtools::install_github('kevinblighe/EnhancedVolcano')

# Clean data for plotting in volcano
volcano.data <- clean.data.volcano(volcano.data, sig.only = FALSE)

# Create vector of top up/down-regulated genes to highlight
goi.up <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(10, logFC)

goi.dn <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(-10, logFC)

# combine vectors
goi <- unique(c(rownames(goi.up), rownames(goi.dn)))

keyvals <- colour.points.volcano(volcano.data,
                                 highlight.selected = TRUE,
                                 selected.cols = c("Green", "Orange"),
                                 highlight.genes = goi)

size.vec <- size.points.volcano(volcano.data,
                                highlight.genes = goi)

alpha.vec <- alpha.points.volcano(volcano.data,
                                  highlight.genes = goi)


custom.enhanced.volcano(volcano.data,
                       selectLab = goi,
                       pointSize = size.vec,
                       colAlpha = alpha.vec,
                       colCustom = keyvals,
                       FCcutoff = 1,
                       title = "Spleen_Pos_vs_Neg_Manuscript")
}
\author{
Dillon Corvino
}
