% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_functions.R
\name{annotate.clinical.data}
\alias{annotate.clinical.data}
\title{Annotate clinical data with gene expression signature}
\usage{
annotate.clinical.data(patients.list, dataset.name)
}
\arguments{
\item{patients.list}{list of patients with high and low gene expression signature}

\item{dataset.name}{name of dataset to retrieve clinical data from}
}
\value{
dataframe of clinical data annotated with gene expression signature
}
\description{
Annotate clinical data with gene expression signature
}
\examples{
annotate.clinical.data(patients.list, dataset.name)
}
