% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_heatmap.R
\name{basic.heatmap}
\alias{basic.heatmap}
\title{Basic Heatmap function}
\usage{
basic.heatmap(
  df,
  Col.cluster = FALSE,
  Row.cluster = FALSE,
  title.var = "",
  colsepvar = NULL,
  row.size = 0.4,
  scale.var = "row",
  dist.method = "euclidean",
  hclust.method = "ward.D"
)
}
\arguments{
\item{df}{dataframe}

\item{Col.cluster}{logical should columns be clustered}

\item{Row.cluster}{logical should rows be clustered}

\item{title.var}{character for what tilte to give the plotted heatmap}

\item{colsepvar}{integer}

\item{row.size}{numeric}

\item{scale.var}{character either "row", "column", or "none"}

\item{dist.method}{character for which distance metric to use available options are "euclidean", "maximum", "manhattan", "canberra", "binary"}

\item{hclust.method}{character for which cluster metric to use available options are "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid"}
}
\value{
heatmap
}
\description{
Heatmap function
}
\examples{
basic.heatmap(df, 
             Col.cluster = FALSE,
             Row.cluster = FALSE,
             title.var = "",
             colsepvar = NULL,
             row.size = 0.4,
             dist.method = "euclidean",
             hclust.method = "ward.D")
}
\author{
Dillon corvino
}
