% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bus_pop_average}
\alias{bus_pop_average}
\title{Business Population Average 2010 - 2022}
\format{
\subsection{\code{bus_pop_average}}{

A data frame
}
}
\usage{
bus_pop_average
}
\description{
Tidy data version of regional business population average from 2010 - 2022
}
\keyword{datasets}
