% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bus_pop_long}
\alias{bus_pop_long}
\title{Business Population Estimates 2010 - 2022}
\format{
\subsection{\code{bus_pop_long}}{

A data frame
}
}
\usage{
bus_pop_long
}
\description{
Tidy data version of regional business population estimates for each year 2010 - 2022
}
\keyword{datasets}
