% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{calc.moving.average}
\alias{calc.moving.average}
\title{Calculate moving average}
\usage{
calc.moving.average(x, n = 1, centered = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{The number of values to average over}

\item{centered}{Whether to center the average}
}
\value{
A vector of the same length as x
}
\description{
function to calculate the moving average, was adapted from Riesenberg, S. et al.Nat. Commun. 2015, doi: 10.1038/ncomms9755
See also http://www.cookbook-r.com/ free usage under creative commons licence http://creativecommons.org/licenses/by-sa/3.0/legalcode
}
\examples{
calc.moving.average(c(1,2,3,4,5), n=3)
}
