% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{clean.data.volcano}
\alias{clean.data.volcano}
\title{Clean data for volcano plot}
\usage{
clean.data.volcano(input.data, sig.only = TRUE)
}
\arguments{
\item{input.data}{data.frame for filtering, should have column FDR and logFC}

\item{sig.only}{logical only keep significant FDR < 0.05 genes}
}
\value{
data.frame which contains only necessary data for volcano plotting
}
\description{
Clean std RNAseq formatted data for plotting in volcano plot
}
\examples{

# "EnhancedVolcano" package from Github repo
# devtools::install_github('kevinblighe/EnhancedVolcano')

# Clean data for plotting in volcano
volcano.data <- clean.data.volcano(volcano.data, sig.only = FALSE)

# Create vector of top up/down-regulated genes to highlight
goi.up <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(10, logFC)

goi.dn <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(-10, logFC)

# combine vectors
goi <- unique(c(rownames(goi.up), rownames(goi.dn)))

keyvals <- colour.points.volcano(volcano.data, 
                                 highlight.selected = TRUE,
                                 selected.cols = c("Green", "Orange"),
                                 highlight.genes = goi)

size.vec <- size.points.volcano(volcano.data, 
                                highlight.genes = goi)

alpha.vec <- alpha.points.volcano(volcano.data, 
                                  highlight.genes = goi)
                        
                        
custom.enhanced.volcano(volcano.data, 
                       selectLab = goi, 
                       pointSize = size.vec, 
                       colAlpha = alpha.vec, 
                       colCustom = keyvals, 
                       FCcutoff = 1, 
                       title = "Spleen_Pos_vs_Neg_Manuscript")
}
\author{
Dillon Corvino
}
