% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{clipboard_it}
\alias{clipboard_it}
\title{Copy data table to the clipboard}
\usage{
clipboard_it(data)
}
\value{
data table to the clipboard
}
\description{
Copy data table to the clipboard
}
\examples{
Will copy the data table output from a script to the clipboard
data \%m\%
 clipboard_it()
}
