% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{co_ho_charges_function}
\alias{co_ho_charges_function}
\title{Companies House Charges Data}
\usage{
co_ho_charges_function(crn_list)
}
\arguments{
\item{crn_list}{is a tibble of companies house reference numbers}
}
\value{
a tibble containing information from the companies house charges API
}
\description{
Companies House Charges Data
}
\examples{
crn_list <- data \%>\%
    distinct(crn)
co_ho_charges_function(crn_list = crn_list)
}
