% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{co_ho_signif_control_function}
\alias{co_ho_signif_control_function}
\title{Companies House Significant Control Data}
\usage{
co_ho_signif_control_function(crn_list)
}
\arguments{
\item{crn_list}{is a tibble of companies house reference numbers}
}
\value{
a tibble containing information from the companies house significant control API
}
\description{
Companies House Significant Control Data
}
\examples{
crn_list <- data \%>\%
    distinct(crn)
co_ho_signif_control_function(crn_list = crn_list)
}
