% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{colour.points.volcano}
\alias{colour.points.volcano}
\title{Colour points in volcano plot}
\usage{
colour.points.volcano(
  volcano.data,
  increase.col = "Red",
  decreased.col = "Blue",
  highlight.selected = FALSE,
  selected.cols = c("Green", "Orange"),
  highlight.genes = goi,
  FDR.cutoff = 0.05,
  logFC.cutoff = 1
)
}
\arguments{
\item{volcano.data}{A dataframe of data formatted for plotting as a volcano plot.}

\item{increase.col}{The colour to use for genes which are increased.}

\item{decreased.col}{The colour to use for genes which are decreased.}

\item{highlight.selected}{A logical value indicating whether to highlight selected genes.}

\item{selected.cols}{A vector of colours to use for selected genes.}

\item{highlight.genes}{A vector of gene names to highlight.}

\item{FDR.cutoff}{The FDR cutoff to use for determining which genes are increased/decreased.}

\item{logFC.cutoff}{The logFC cutoff to use for determining which genes are increased/decreased.}
}
\value{
A vector of colours for each gene.
}
\description{
This function takes a dataframe of volcano plot data and returns a vector of colours for each gene.
}
\examples{

# "EnhancedVolcano" package from Github repo
# devtools::install_github('kevinblighe/EnhancedVolcano')

# Clean data for plotting in volcano
volcano.data <- clean.data.volcano(volcano.data, sig.only = FALSE)

# Create vector of top up/down-regulated genes to highlight
goi.up <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(10, logFC)

goi.dn <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(-10, logFC)

# combine vectors
goi <- unique(c(rownames(goi.up), rownames(goi.dn)))

keyvals <- colour.points.volcano(volcano.data,
                                 highlight.selected = TRUE,
                                 selected.cols = c("Green", "Orange"),
                                 highlight.genes = goi)

size.vec <- size.points.volcano(volcano.data,
                                highlight.genes = goi)

alpha.vec <- alpha.points.volcano(volcano.data,
                                  highlight.genes = goi)


custom.enhanced.volcano(volcano.data,
                       selectLab = goi,
                       pointSize = size.vec,
                       colAlpha = alpha.vec,
                       colCustom = keyvals,
                       FCcutoff = 1,
                       title = "Spleen_Pos_vs_Neg_Manuscript")
}
\author{
Dillon Corvino
}
