% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Gene_IDs.R
\name{convert.human.to.mouse}
\alias{convert.human.to.mouse}
\title{Convert Human genes to mouse genes}
\usage{
convert.human.to.mouse(x = c("GZMA", "GZMB", "XCL1"))
}
\arguments{
\item{genes}{a character vector of human gene names}

\item{use.jax}{logic variable to convert genes using data from jax website, to use incase connection error with ensemble}
}
\value{
a data frame with mouse and human gene names
}
\description{
Convert Human genes to mouse genes
}
\examples{
convert.human.to.mouse(genes = c("GZMA", "GZMB", "XCL1"))

}
\author{
written by radiaj in R bloggers, Oct 14, 2016 https://www.r-bloggers.com/2016/10/converting-mouse-to-human-gene-names-with-biomart-package/
}
