% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Gene_IDs.R
\name{convert.mouse.to.human}
\alias{convert.mouse.to.human}
\title{Convert Mouse genes to human genes}
\usage{
convert.mouse.to.human(mouse.genes)
}
\arguments{
\item{x}{a character vector of mouse gene names}
}
\value{
a data frame with mouse and human gene names
}
\description{
Convert Mouse genes to human genes
}
\examples{
convert.mouse.to.human(c("Gzma", "Gzmb", "Xcl1"))

}
\author{
written by radiaj in R bloggers, Oct 14, 2016 https://www.r-bloggers.com/2016/10/converting-mouse-to-human-gene-names-with-biomart-package/
}
