% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core}
\alias{core}
\title{Old Core Definitions}
\format{
\subsection{\code{core}}{

A character vector
}
}
\usage{
core
}
\description{
Old Core Definitions useful for grouping IUK Product Types together
}
\keyword{datasets}
