% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_correlation.R
\name{correlation.heatmaps}
\alias{correlation.heatmaps}
\title{Correlation.heatmaps}
\usage{
correlation.heatmaps(
  seurat.object,
  correlation.df,
  up.n = 20,
  dn.n = 20,
  goi = "NKG7",
  output.dir = "results/figures/Correlation_analysis/"
)
}
\arguments{
\item{seurat.object}{Seurat object, should be downsampled and idents and assay set to those of interest}

\item{correlation.df}{Dataframe of correlation values, output from sc.correlation function}

\item{up.n}{Number of up regulated genes to show}

\item{dn.n}{Number of downregulated genes to show}

\item{goi}{Gene of interest - used in calculating correlation}

\item{output.dir}{Output directory for saving figures}
}
\value{
print and export heatmaps with batlow.pal colour scheme
}
\description{
Creates heatmaps of the top correlated genes to a gene of interest, output from sc.correlation function
}
\examples{


DefaultAssay(downsampled.seurat) <- "SCT"
Idents(downsampled.seurat) <- downsampled.seurat@meta.data$celltype.l3

small.seurat <- subset(downsampled.seurat,
                      downsample = 100)

small.seurat <- ScaleData(small.seurat)

correlation.heatmaps(seurat.object = small.seurat,
                    correlation.df = cor.df,
                    up.n = 20,
                    dn.n = 20,
                    goi = "NKG7",
                    output.dir = "results/figures/Correlation_analysis/")
}
