% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{custom.enhanced.volcano}
\alias{custom.enhanced.volcano}
\title{Custom default settings for enhanced volcano plot}
\usage{
custom.enhanced.volcano(
  toptable,
  lab = rownames(toptable),
  x = "logFC",
  y = "FDR",
  selectLab = NULL,
  xlim = c(min(toptable[[x]], na.rm = TRUE) - 1.5, max(toptable[[x]], na.rm = TRUE) +
    1.5),
  ylim = c(0, max(-log10(toptable[[y]]), na.rm = TRUE) + 5),
  xlab = bquote(~Log[2] ~ "fold change"),
  ylab = bquote(~-Log[10] ~ adjusted ~ italic(P)),
  axisLabSize = 18,
  title = "Volcano plot",
  subtitle = "",
  caption = paste0("total = ", nrow(toptable), " variables"),
  titleLabSize = 18,
  subtitleLabSize = 14,
  captionLabSize = 14,
  pCutoff = 0.05,
  FCcutoff = 0.25,
  cutoffLineType = "longdash",
  cutoffLineCol = "black",
  cutoffLineWidth = 0.4,
  pointSize = 5,
  labSize = 4,
  labCol = "black",
  labFace = "plain",
  labhjust = 0.5,
  labvjust = 1.5,
  boxedLabels = FALSE,
  shape = 19,
  shapeCustom = NULL,
  col = c("grey30", "forestgreen", "royalblue", "red2"),
  colCustom = NULL,
  colAlpha = 0.2,
  colGradient = NULL,
  colGradientBreaks = c(pCutoff, 1),
  colGradientLabels = c("0", "1.0"),
  colGradientLimits = c(0, 1),
  legendLabels = c("NS", expression(Log[2] ~ FC), "p-value", expression(p - value ~ and ~
    log[2] ~ FC)),
  legendPosition = "top",
  legendLabSize = 14,
  legendIconSize = 5,
  encircle = NULL,
  encircleCol = "black",
  encircleFill = "pink",
  encircleAlpha = 3/4,
  encircleSize = 2.5,
  shade = NULL,
  shadeFill = "grey",
  shadeAlpha = 1/2,
  shadeSize = 0.01,
  shadeBins = 2,
  drawConnectors = TRUE,
  widthConnectors = 0.5,
  typeConnectors = "open",
  endsConnectors = "first",
  lengthConnectors = unit(0.01, "npc"),
  colConnectors = "grey10",
  arrowheads = TRUE,
  hline = NULL,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = NULL,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  gridlines.major = TRUE,
  gridlines.minor = FALSE,
  border = "partial",
  borderWidth = 0.8,
  borderColour = "black"
)
}
\arguments{
\item{toptable}{A data.frame with the columns 'logFC' and 'FDR'}

\item{lab}{A character vector of labels for the points}

\item{x}{A character vector of the column name for the x-axis}

\item{y}{A character vector of the column name for the y-axis}

\item{selectLab}{A character vector of labels to be selected}

\item{xlim}{A numeric vector of length 2 for the x-axis limits}

\item{ylim}{A numeric vector of length 2 for the y-axis limits}

\item{xlab}{A character vector of the x-axis label}

\item{ylab}{A character vector of the y-axis label}

\item{axisLabSize}{A numeric value for the axis label size}

\item{title}{A character vector of the plot title}

\item{subtitle}{A character vector of the plot subtitle}

\item{caption}{A character vector of the plot caption}

\item{titleLabSize}{A numeric value for the title label size}

\item{subtitleLabSize}{A numeric value for the subtitle label size}

\item{captionLabSize}{A numeric value for the caption label size}

\item{pCutoff}{A numeric value for the p-value cutoff}

\item{FCcutoff}{A numeric value for the fold change cutoff}

\item{cutoffLineType}{A character vector of the cutoff line type}

\item{cutoffLineCol}{A character vector of the cutoff line color}

\item{cutoffLineWidth}{A numeric value for the cutoff line width}

\item{pointSize}{A numeric value for the point size}

\item{labSize}{A numeric value for the label size}

\item{labCol}{A character vector of the label color}

\item{labFace}{A character vector of the label face}

\item{labhjust}{A numeric value for the label horizontal justification}

\item{labvjust}{A numeric value for the label vertical justification}

\item{boxedLabels}{A logical value for boxed labels}

\item{shape}{A numeric value for the point shape}

\item{shapeCustom}{A numeric vector of the custom point shape}

\item{col}{A character vector of the point color}

\item{colCustom}{A character vector of the custom point color}

\item{colAlpha}{A numeric value for the point color alpha}

\item{colGradient}{A character vector of the point color gradient}

\item{colGradientBreaks}{A numeric vector of the point color gradient breaks}

\item{colGradientLabels}{A character vector of the point color gradient labels}

\item{colGradientLimits}{A numeric vector of the point color gradient limits}

\item{legendLabels}{A character vector of the legend labels}

\item{legendPosition}{A character vector of the legend position}

\item{legendLabSize}{A numeric value for the legend label size}

\item{legendIconSize}{A numeric value for the legend icon size}

\item{encircle}{A character vector of the labels to be encircled}

\item{encircleCol}{A character vector of the encircle color}

\item{encircleFill}{A character vector of the encircle fill}

\item{encircleAlpha}{A numeric value for the encircle alpha}

\item{encircleSize}{A numeric value for the encircle size}

\item{shade}{A character vector of the labels to be shaded}

\item{shadeFill}{A character vector of the shade fill}

\item{shadeAlpha}{A numeric value for the shade alpha}

\item{shadeSize}{A numeric value for the shade size}

\item{shadeBins}{A numeric value for the shade bins}

\item{drawConnectors}{A logical value for drawing connectors}

\item{widthConnectors}{A numeric value for the connector width}

\item{typeConnectors}{A character vector of the connector type}

\item{endsConnectors}{A character vector of the connector ends}

\item{lengthConnectors}{A numeric value for the connector length}

\item{colConnectors}{A character vector of the connector color}

\item{arrowheads}{A logical value for arrowheads}

\item{hline}{A numeric value for the horizontal line}

\item{hlineType}{A character vector of the horizontal line type}

\item{hlineCol}{A character vector of the horizontal line color}

\item{hlineWidth}{A numeric value for the horizontal line width}

\item{vline}{A numeric value for the vertical line}

\item{vlineType}{A character vector of the vertical line type}

\item{vlineCol}{A character vector of the vertical line color}

\item{vlineWidth}{A numeric value for the vertical line width}

\item{gridlines.major}{A logical value for major gridlines}

\item{gridlines.minor}{A logical value for minor gridlines}

\item{border}{A character vector of the border}

\item{borderWidth}{A numeric value for the border width}

\item{borderColour}{A character vector of the border color}
}
\value{
A ggplot object
}
\description{
Enhanced volcano plot
}
\examples{

# "EnhancedVolcano" package from Github repo
# devtools::install_github('kevinblighe/EnhancedVolcano')

# Clean data for plotting in volcano
volcano.data <- clean.data.volcano(volcano.data, sig.only = FALSE)

# Create vector of top up/down-regulated genes to highlight
goi.up <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(10, logFC)

goi.dn <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(-10, logFC)

# combine vectors
goi <- unique(c(rownames(goi.up), rownames(goi.dn)))

keyvals <- colour.points.volcano(volcano.data, 
                                 highlight.selected = TRUE,
                                 selected.cols = c("Green", "Orange"),
                                 highlight.genes = goi)

size.vec <- size.points.volcano(volcano.data, 
                                highlight.genes = goi)

alpha.vec <- alpha.points.volcano(volcano.data, 
                                  highlight.genes = goi)
                        
                        
custom.enhanced.volcano(volcano.data, 
                       selectLab = goi, 
                       pointSize = size.vec, 
                       colAlpha = alpha.vec, 
                       colCustom = keyvals, 
                       FCcutoff = 1, 
                       title = "Spleen_Pos_vs_Neg_Manuscript")
}
\author{
Kevin Blighe, Sharmila Rana, Myles Lewis, authors of EnhancedVolcano package
}
