% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{dataWarehouse_function_specific}
\alias{dataWarehouse_function_specific}
\title{Data Warehouse Function}
\usage{
dataWarehouse_function_specific(table_choice = "Tables")
}
\arguments{
\item{table_choice}{is the data warehouse table we want to return, leaving this blank will return the available tables in the console}
}
\value{
a data frame
}
\description{
Data Warehouse Function
}
\examples{
dataWarehouse_function_specific("Application Summaries")
}
