% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dendrogram.R
\name{dendrogram.samples}
\alias{dendrogram.samples}
\title{Dendrogram function}
\usage{
dendrogram.samples(
  data.matrix,
  hc_metric.var = "euclidean",
  hc_method.var = "ward.D2",
  type.var = NULL,
  output.dir = "results/figures/"
)
}
\arguments{
\item{data.matrix}{A matrix of data}

\item{hc_metric.var}{A string of the distance metric to be used in the hierarchical clustering options include "euclidean", "maximum", "manhattan", "canberra", "binary"}

\item{hc_method.var}{A string of the clustering method to be used in the hierarchical clustering options include "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{type.var}{A string of the type of dendrogram to be plotted options include "phylogram”, “cladogram”, “fan”, “unrooted”, “radial”, if null, standard dendrogram plotted}

\item{output.dir}{A string of the output directory}
}
\value{
saves a pdf file of a dendrogram
}
\description{
Dendrogram function
}
\examples{
dendrogram.samples(data.matrix = data.matrix, 
                   hc_metric.var = "euclidean",
                   hc_method.var = "ward.D2", 
                   type.var = "unrooted",
                   output.dir = "results/figures/")
}
