% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_functions.R
\name{download.TCGA.RNAseq}
\alias{download.TCGA.RNAseq}
\title{Download RNAseq data from TCGA}
\usage{
download.TCGA.RNAseq(
  dataset.name,
  raw.data.dir = "GDCdata/",
  rda.dir = "data/"
)
}
\arguments{
\item{dataset.name}{the name of the dataset to download e.g TCGA-GBM, TCGA-HNSC}

\item{raw.data.dir}{the directory to save raw data}

\item{rda.dir}{the directory to save the processed R compatiable data to}
}
\value{
counts table of raw RNAseq values
}
\description{
This helper function downloads RNAseq data for the indicated dataset.
The raw data will be stored in a directory "raw.data.dir" and an rda version will be saved in 'rda.dir'
}
\examples{
download.TCGA.RNAseq("TCGA-BRCA")
}
