% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyFunction.R
\name{fy_function}
\alias{fy_function}
\title{Fiscal Year Function}
\usage{
fy_function(.data, date_field)
}
\arguments{
\item{.data}{is a dataframe}

\item{date_field}{is a date column in the dataframe}
}
\value{
an additional column in the dataframe with the correct fiscal year
}
\description{
Fiscal Year Function
}
\examples{
data_With_fy <- data_frame \%>\% 
  mutate(fy_function(.,date_field = To))
}
