% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colour_scheme.R
\name{Get.batlow}
\alias{Get.batlow}
\title{Get Batlow}
\usage{
Get.batlow()
}
\value{
a vector of colors from the batlow palette
}
\description{
This function returns a vector of colors from the batlow palette
}
\examples{
# requires scico package
# install.packages("scico")
# library("scico")
batlow.pal <- Get.batlow()
}
