% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janus_function.R
\name{janus_function_specific}
\alias{janus_function_specific}
\title{Janus Function}
\usage{
janus_function_specific(table_choice = "Tables")
}
\arguments{
\item{table_choice}{is the Janus equivalent data warehouse table to return, leaving this blank will return the available tables in the console}
}
\value{
a data frame
}
\description{
Janus Function
}
\examples{
janus_function_specific("Applications")
}
