% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{lm_summary_table_function}
\alias{lm_summary_table_function}
\title{lm() summary dataframe function}
\usage{
lm_summary_table_function(summary_table)
}
\value{
a tibble of releavtn output from the summary function for a linear model
}
\description{
lm() summary dataframe function
}
\examples{
Requires a lm model
summary_table <- summary(linear_model) \%>\% 
 lm_summary_table_function()
}
