% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparisons.R
\name{make_comparisons}
\alias{make_comparisons}
\title{Make vecotr of comparisons for EdgeR}
\usage{
make_comparisons(grouper)
}
\arguments{
\item{grouper}{vector to generate comaprisons from}
}
\value{
vector of comparisons for use with edgeR
}
\description{
Generate a vector of all comparisons in a list for edgeR from a vector of all
qualities to compare by, seperated by an underscore
}
\details{
Proud of this!
MakeComparisons()
}
\keyword{EdgeR}
