% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent_colour.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent}
\usage{
makeTransparent(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{a color e.g "red"}

\item{percent}{a number between 0 and 100 indicating percentage transparent}

\item{name}{an optional name for the color}
}
\value{
a new color with transparency
}
\description{
Make a color transparent
}
\examples{
makeTransparent(color = "red", percent = 50, name = NULL)

}
\author{
Ricardo Oliveros-Ramos https://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color
}
