% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving.average}
\alias{moving.average}
\title{Moving average plot}
\usage{
moving.average(
  tcga.datasets = c("laml_tcga", "luad_tcga"),
  output.dir = "results/moving_average/",
  primary.goi = "TLR2",
  goi = c("TLR6", "CSF1R", "CD14", "CD8B", "NCAM1", "IFNG", "GZMB"),
  sample.width = 20,
  laxis.col = "blue"
)
}
\arguments{
\item{tcga.datasets}{A vector of TCGA datasets to use for the analysis.}

\item{output.dir}{The directory to save the output files.}

\item{primary.goi}{The primary GOI to use for ranking the samples.}

\item{goi}{A vector of GOI to plot against the primary GOI.}

\item{sample.width}{The sample window size for the moving average calculation.}

\item{laxis.col}{The colour of the dependent gene.}
}
\value{
A plot of the moving average of a gene of interest (GOI) against the expression of another GOI.
}
\description{
This function plots the moving average of a gene of interest (GOI) against the expression of another GOI.
}
\examples{

moving.average()
}
