% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{namesFunction}
\alias{namesFunction}
\title{Name Repair Function}
\usage{
namesFunction(nms)
}
\arguments{
\item{nms}{used only in the .name_repair argument of readxl::read_excel and without parentheses}
}
\value{
variable names in dataframe in upper camel
}
\description{
Name Repair Function
}
\examples{
output <- readxl::read_excel("C:/data.xlsx",
                             .name_repair = namesFunction)
}
