% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseUniNames.R
\name{normalise_university_name_function}
\alias{normalise_university_name_function}
\title{Normalise Univeristy Names Function}
\usage{
normalise_university_name_function(uni_name_list)
}
\arguments{
\item{uni_name_list}{is a tibble with the first column university names or an atomic vector}
}
\value{
a tibble called noramlised_uni_name_list with original and normalised names or a mutated atomic vector
}
\description{
Normalise Univeristy Names Function
}
\examples{
for tibble
uni_name_list <- trans_data \%>\% 
  filter(EnterpriseSizeClean == "Academic") \%>\% 
  distinct(ParticipantName)
  
  normalise_university_name_function(uni_name_list)
  
  or atomic vector
  trans_data \%>\%
    filter(EnterpriseSize == "Academic") \%>\% 
    mutate(UniversityName = normalise_university_name_function(ParticipantName))
}
