% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_correlation.R
\name{plot.cor.data}
\alias{plot.cor.data}
\title{Plot correlation data}
\usage{
\method{plot}{cor.data}(cor.df, output.dir = "output/figures/", goi = "NKG7", top.n.val = 30)
}
\arguments{
\item{cor.df}{dataframe of correlation values}

\item{output.dir}{output directory}

\item{goi}{gene of interest}

\item{top.n.val}{number of top values to plot}
}
\value{
plot of correlation analysis
}
\description{
Plot correlation data
}
\examples{
plot.cor.data(cor.df = cor.df,
             output.dir = "output/figures/",
             goi = "NKG7",
             top.n.val = 30)
}
