% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_functions.R
\name{plot.OS}
\alias{plot.OS}
\title{plot the overall survival}
\usage{
\method{plot}{OS}(
  annotated.clin.data,
  dataset.name,
  goi,
  output.dir = "figures/",
  with.risk = TRUE,
  with.conf = TRUE
)
}
\arguments{
\item{annotated.clin.data}{output from annotate.clinical.data function}

\item{dataset.name}{name of TCGA dataset}

\item{output.dir}{file.path where plots should be saved}

\item{with.risk}{logical, should risk table be included}

\item{with.conf}{logical should confidence interval be included}
}
\value{
saves an OS plot to the specified location
}
\description{
Generate Kaplan-Meier Overall Survival plot
}
