% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{postcode_function}
\alias{postcode_function}
\title{Additional geo data from postcodes}
\usage{
postcode_function(post_code_list)
}
\arguments{
\item{post_code_list}{is a tibble postcodes}
}
\value{
a tibble containing additional geographic data from the postcodes.io API
}
\description{
Additional geo data from postcodes
}
\examples{
post_code_list <- data \%>\%
    distinct(postcode)
postcode_function(post_code_list = post_code_list)
}
