% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print_function}
\alias{print_function}
\title{print function}
\usage{
print_function(input = TRUE)
}
\arguments{
\item{input}{This is the input, it accepts true or false values}
}
\value{
prints some text
}
\description{
print function
}
\examples{
print_function(input = TRUE)
print_function(input = FALSE)


}
