% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print_function_2}
\alias{print_function_2}
\title{print function 2}
\usage{
print_function_2(input)
}
\arguments{
\item{input}{takes a number either 1 or 2 or other}
}
\value{
prints some text
}
\description{
print function 2
}
\examples{
print_function_2(1)
print_function_2(input = 2)



}
