% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_correlation.R
\name{probeCor}
\alias{probeCor}
\title{probe correlation}
\usage{
probeCor(em, gp, method = "pearson")
}
\arguments{
\item{em}{a matrix of expression data}

\item{gp}{string - gene of interest to prob correlation against}

\item{method}{a string}
}
\value{
a correlation matrix
}
\description{
calculate correlation between goi and data
}
\examples{
cor.res <- probeCor(filt.data,
                    gp = "NKG7",
                    method = "pearson")

}
\author{
Adapted from Hoelzel code
}
