% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_correlation.R
\name{probeCor.test}
\alias{probeCor.test}
\title{Prob corelation test}
\usage{
probeCor.test(em, gp, method = "pearson")
}
\arguments{
\item{em}{a matrix of expression data}

\item{gp}{string - gene of interest to prob correlation against}

\item{method}{a string of method used to calculate the correlation}
}
\value{
a list of correlation test result
}
\description{
This function is used to calculate the statistical significance of correlation
}
\examples{
cor.res.pval <- probeCor.test(filt.data,
                             gp = "NKG7",
                             method = "pearson")
}
\author{
Adapted from Hoelzel code
}
