% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seurat.R
\name{process_seurat}
\alias{process_seurat}
\title{Process Seurat objects}
\usage{
process_seurat(dirlist, outlist)
}
\arguments{
\item{dirlist}{takes a list of 10x sequencing directories to process}

\item{outlist}{takes a list of what you would like the output objects to be called}
}
\value{
a list of Seurat objects that in which doublets and dead cells have been filtered
variable features have been identified, normalized by SCtransform,
}
\description{
This command creates a Seurat object, filters for doublets and dead cells
runs UMap, Tsne, and creates a clusterplot
}
\details{
you may need to modify the criteria for subsetting. If your elbow plot is far from
18 you will need to change it most likely.
process_seurat()
}
\keyword{Seurat}
\keyword{processing}
