% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_groups}
\alias{product_groups}
\title{Product Grouping Definitions}
\format{
\subsection{\code{product_groups}}{

A character vector
}
}
\usage{
product_groups
}
\description{
New Product Grouping useful for grouping IUK Product Types together
}
\keyword{datasets}
