% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_correlation.R
\name{sc.correlation}
\alias{sc.correlation}
\title{Correlation analysis}
\usage{
sc.correlation(
  data.slot,
  goi = "NKG7",
  remove.quantile = 0.3,
  output.tables = "results/tables/Correlation_analysis/"
)
}
\arguments{
\item{data.slot}{dataframe from @data slot from a normalised and downsampled seurat object - already downsampled according to ident of interest to a reasonable cell number}

\item{goi}{gene to test correlation against}

\item{remove.quantile}{lowest X\% expressed genes in dataset to remove from analysis}

\item{output.tables}{output directory for correlation results}
}
\value{
correlation dataframe and saves a copy to designated output location
}
\description{
Calculate correlation between a gene of interest and all other genes in the dataset
}
\examples{


# Set Idents
Idents(seurat.object) <- seurat.object@meta.data$cluster
# Downsample object as necessary
# extract normalised data table
normed.data <- as.matrix(seurat.object@assays$RNA@data)

# calculate correlation
sc.correlation(data.slot = normed.data,
              goi = "NKG7",
              remove.quantile = 0.3,
              output.tables = "results/tables/Correlation_analysis/")
}
