% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{size.points.volcano}
\alias{size.points.volcano}
\title{Size points}
\usage{
size.points.volcano(
  input.data,
  highlight.genes = "NKG7",
  baseline.size = 5,
  highlight.size = 5
)
}
\arguments{
\item{input.data}{A data frame with gene expression data}

\item{highlight.genes}{A vector of gene names to be highlighted}

\item{baseline.size}{The size of points for genes not in highlight.genes}

\item{highlight.size}{The size of points for genes in highlight.genes}
}
\value{
A vector of point sizes
}
\description{
This function is used to set the size of points in a plot
}
\examples{

# "EnhancedVolcano" package from Github repo
# devtools::install_github('kevinblighe/EnhancedVolcano')

# Clean data for plotting in volcano
volcano.data <- clean.data.volcano(volcano.data, sig.only = FALSE)

# Create vector of top up/down-regulated genes to highlight
goi.up <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(10, logFC)

goi.dn <- volcano.data \%>\%
dplyr::filter(FDR < 0.05) \%>\%
top_n(-10, logFC)

# combine vectors
goi <- unique(c(rownames(goi.up), rownames(goi.dn)))

keyvals <- colour.points.volcano(volcano.data,
                                 highlight.selected = TRUE,
                                 selected.cols = c("Green", "Orange"),
                                 highlight.genes = goi)

size.vec <- size.points.volcano(volcano.data,
                                highlight.genes = goi)

alpha.vec <- alpha.points.volcano(volcano.data,
                                  highlight.genes = goi)


custom.enhanced.volcano(volcano.data,
                       selectLab = goi,
                       pointSize = size.vec,
                       colAlpha = alpha.vec,
                       colCustom = keyvals,
                       FCcutoff = 1,
                       title = "Spleen_Pos_vs_Neg_Manuscript")
}
\author{
Dillon Corvino
}
