% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smes}
\alias{smes}
\title{SME definitions}
\format{
\subsection{\code{smes}}{

A character vector with 3 entries
}
}
\usage{
smes
}
\description{
SME definitions useful for grouping organisations together
}
\keyword{datasets}
