% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{survey_id_function}
\alias{survey_id_function}
\title{Qualtrics Survey IDs}
\usage{
survey_id_function()
}
\value{
a tibble of survey ids and associated names
}
\description{
Qualtrics Survey IDs
}
\examples{
survey_id_function()
}
