% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{survey_questions_function}
\alias{survey_questions_function}
\title{Qualtrics Survey Questions}
\usage{
survey_questions_function()
}
\value{
a tibble of survey questions
}
\description{
Qualtrics Survey Questions
}
\examples{
survey_response_function()
}
