% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{survey_response_function}
\alias{survey_response_function}
\title{Qualtrics Survey Responses}
\usage{
survey_response_function()
}
\value{
a tibble of survey responses
}
\description{
Qualtrics Survey Responses
}
\examples{
survey_response_function()
}
