% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_functions.R
\name{TCGA.OS}
\alias{TCGA.OS}
\title{Plot overall survival curve for a selected TCGA dataset based on gene signature}
\usage{
TCGA.OS(
  TCGA.dataset,
  gene.signature,
  gene.name = NULL,
  rda.file.path = NULL,
  plot.dir = "figures/",
  data.dir = "data/"
)
}
\arguments{
\item{TCGA.dataset}{A character string indicating the TCGA dataset to be used e.g "TCGA-HNSC"}

\item{gene.signature}{A character string indicating the gene of interest or vector of strings indicating gene signature}

\item{gene.name}{A character string indicating the name of the gene of interest}

\item{rda.file.path}{A character string indicating the path to the rda file containing the raw data}

\item{plot.dir}{File path for where to save OS plot}

\item{data.dir}{File path for where to save data to}
}
\value{
A plot of overall survival is saved to the desired output location
}
\description{
This function will download TCGA RNAseq data, process it, threshold patients based on provided gene signature
then plot an overall surival curve
}
\details{
gene.name param allows the user to define a different name if the gene id is an uncharacteristic name.
For example, gene.signature = "CD226" but perhaps you want plots to be named "DNAM1" therefore set gene.name = "DNAM1"
If no gene.name is set, it automatically inherits gene.signature

Provide a rda.file.path if data has already been downloaded. this will save signficant time
}
\examples{
TCGA.OS(TCGA.dataset = "TCGA-HNSC",
        gene.signature = "NKG7",
        rda.file.path = "TCGA-HNSC_Transcriptome_Profiling_raw_counts.rda")
}
