% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_functions.R
\name{threshold.goi}
\alias{threshold.goi}
\title{threshold.goi}
\usage{
threshold.goi(input.data, goi)
}
\arguments{
\item{input.data}{A matrix of gene expression data.}

\item{goi}{A gene of interest.}
}
\value{
A list of patients with high and low expression of the goi.
}
\description{
This function takes a gene of interest (goi) and thresholds the expression data to identify patients with high and low expression of the goi.
}
\examples{
threshold.goi(input.data = expression.data, goi = "NKG7")
}
