% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{transDataFunction}
\alias{transDataFunction}
\title{Transparency Data Function}
\usage{
transDataFunction()
}
\value{
the latest version of our published transparency data set
}
\description{
Transparency Data Function
}
\examples{
transDataFunction()
}
