% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_optimisation.R
\name{UMAP.optimise}
\alias{UMAP.optimise}
\title{UMAP optimisation function}
\usage{
UMAP.optimise(
  input.seurat,
  output.dir = "results/optimising_UMAP/",
  init.min.dist = 0.001,
  init.neigh = 5,
  min.dist.step = 2,
  neigh.step = 20
)
}
\arguments{
\item{input.seurat}{A Seurat object}

\item{output.dir}{A directory to save the output to. If NULL, will create a directory called "output/Optimising_UMAP"}

\item{init.min.dist}{The starting min.dist value. sensible values = 0.001 - 0.5}

\item{init.neigh}{The starting n.neighbor value. sensible values = 5 - 50}

\item{min.dist.step}{The step size for min.dist. Default = 2}

\item{neigh.step}{The step size for n.neighbor. Default = 20}
}
\value{
Numerous .pdf files are exported and saved to the desired output.dir
}
\description{
This function will calculate UMAP for a range of min.dist and n.neighbor values.
}
\examples{
UMAP.optimise(input.seurat = pbmc_small,
             output.dir = NULL,
             init.min.dist = 0.001,
             init.neigh = 5,
             min.dist.step = 2,
             neigh.step = 20)
}
