% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{update_useful_functions}
\alias{update_useful_functions}
\title{update usefulfunctions}
\usage{
update_useful_functions()
}
\value{
the lastest version of the package installed from github
}
\description{
update usefulfunctions
}
\examples{
Will overwrite current installed version but will not update out of date dependancies
update_useful_functions()
}
