% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulfunctions-package.R
\docType{package}
\name{usefulfunctions-package}
\alias{usefulfunctions}
\alias{usefulfunctions-package}
\title{usefulfunctions: Package holds bespoke functions and wrappers developed by Dillon Corvino}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Eomesodermin/usefulfunctions}
  \item Report bugs at \url{https://github.com/Eomesodermin/usefulfunctions/issues}
}

}
\author{
\strong{Maintainer}: Dillon Corvino \email{dilloncorvino@me.com} (\href{https://orcid.org/0000-0003-0683-1369}{ORCID})

}
\keyword{internal}
