% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crnCleanse.R
\name{clean_crn_function}
\alias{clean_crn_function}
\title{CRN Cleansing Fucntion}
\usage{
clean_crn_function(crn_list)
}
\arguments{
\item{crn_list}{is a tibble where the first column is crns}
}
\value{
a tibble with original and cleansed crns
}
\description{
CRN Cleansing Fucntion
}
\examples{
crn_list <- trans_data \%>\% 
    distinct(Crn)
clean_crn_function(crn_list)
}
