% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crnCleanse.R
\name{clean_crn_function}
\alias{clean_crn_function}
\title{CRN Cleansing Fucntion}
\usage{
clean_crn_function(crn_list)
}
\arguments{
\item{crn_list}{can be a tibble where the first column is crns or an atomic vector}
}
\value{
a tibble with original and cleansed crns
}
\description{
CRN Cleansing Fucntion
}
\examples{
as tibble
crn_list <- trans_data \%>\% 
    distinct(Crn)
clean_crn_function(crn_list)

or atomic vector 
clean_crn_trans_data <- clean_crn_trans_data \%>\% 
    mutate(clean_crn = clean_crn_function(Crn)) 

}
