% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulFunctionsPackage.R
\name{co_ho_appointments_function}
\alias{co_ho_appointments_function}
\title{Companies House Appointments}
\usage{
co_ho_appointments_function(officer_list)
}
\value{
a tibble of appointments for a selection of officers
}
\description{
Companies House Appointments
}
