% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Gene_IDs.R
\name{convert.human.to.mouse}
\alias{convert.human.to.mouse}
\title{Convert Human genes to mouse genes}
\usage{
convert.human.to.mouse(human.genes)
}
\arguments{
\item{x}{a character vector of human gene names}
}
\value{
a data frame with mouse and human gene names
}
\description{
Convert Human genes to mouse genes
}
\examples{
convert.human.to.mouse(c("GZMA", "GZMB", "XCL1"))

}
\author{
written by radiaj in R bloggers, Oct 14, 2016 https://www.r-bloggers.com/2016/10/converting-mouse-to-human-gene-names-with-biomart-package/
}
