% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent_colour.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make a color transparent}
\usage{
makeTransparent("red", "blue", alpha=0.8)
}
\arguments{
\item{...}{a color}

\item{alpha}{a number between 0 and 1}
}
\value{
a color
}
\description{
Make a color transparent
}
\examples{
makeTransparent("red", alpha=0.5)
makeTransparent(rgb(0,0,0,0.5))
}
\author{
Ricardo Oliveros-Ramos https://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color
}
