% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseNniNames.R
\name{normalise_university_name_function}
\alias{normalise_university_name_function}
\title{Normalise Univeristy Names Function}
\usage{
normalise_university_name_function(uni_name_list)
}
\arguments{
\item{uni_name_list}{is a tibble with the first column university names}
}
\value{
a tibble called noramlised_uni_name_list with original and normalised names
}
\description{
Normalise Univeristy Names Function
}
\examples{
uni_name_list <- trans_data \%>\% 
  filter(EnterpriseSizeClean == "Academic") \%>\% 
  distinct(ParticipantName)
  
  normalise_university_name_function(uni_name_list)
}
